requirements.txt
python-telegram-bot>=22.7
python-dotenv==1.0.0
chromadb==0.4.24
numpy==1.26.4
google-genai
ollama>=0.3.0
groq
langgraph==0.1.17
httpx>=0.27.0
psutil
pyotp>=2.9.0
ddgs
aiosqlite==0.20.0
aiofiles==23.2.1
pypdf>=4.0.0
python-docx
pdfplumber>=0.11.0
pandas>=2.2.0
trafilatura>=1.9.0
yfinance
pytest>=8.0.0
pytest-asyncio>=0.23.0
